+++ management/organization_test.go
@@ -454,6 +454,69 @@ func TestOrganizationManager_MemberRoles(t *testing.T) {
 	assert.Len(t, roles.Roles, 0)
 }
 
+func TestOrganizationManager_IsAppEntitlementActive(t *testing.T) {
+	configureHTTPTestRecordings(t)
+
+	org := &Organization{
+		Name:                   auth0.String(fmt.Sprintf("test-organization%v", rand.Intn(999))),
+		DisplayName:            auth0.String("Test Organization"),
+		IsAppEntitlementActive: auth0.Bool(true),
+	}
+
+	err := api.Organization.Create(context.Background(), org)
+	require.NoError(t, err)
+	t.Cleanup(func() {
+		cleanupOrganization(t, org.GetID())
+	})
+	assert.True(t, org.GetIsAppEntitlementActive())
+
+	actualOrg, err := api.Organization.Read(context.Background(), org.GetID())
+	require.NoError(t, err)
+	assert.True(t, actualOrg.GetIsAppEntitlementActive())
+
+	err = api.Organization.Update(context.Background(), org.GetID(), &Organization{IsAppEntitlementActive: auth0.Bool(false)})
+	require.NoError(t, err)
+
+	actualOrg, err = api.Organization.Read(context.Background(), org.GetID())
+	require.NoError(t, err)
+	assert.False(t, actualOrg.GetIsAppEntitlementActive())
+}
+
+func TestOrganizationManager_Clients(t *testing.T) {
+	configureHTTPTestRecordings(t)
+
+	org := givenAnOrganization(t)
+	client := givenAClient(t)
+
+	err := api.Organization.AssociateClient(context.Background(), org.GetID(), client.GetClientID(), true)
+	require.NoError(t, err)
+
+	associatedClient, err := api.Organization.Client(context.Background(), org.GetID(), client.GetClientID())
+	require.NoError(t, err)
+	assert.Equal(t, client.GetClientID(), associatedClient.GetClientID())
+	assert.True(t, associatedClient.GetUseForMemberAccess())
+	assert.Equal(t, client.GetName(), associatedClient.GetClient().GetName())
+
+	clients, err := api.Organization.Clients(context.Background(), org.GetID())
+	require.NoError(t, err)
+	assert.Len(t, clients.Clients, 1)
+	assert.Equal(t, client.GetClientID(), clients.Clients[0].GetClientID())
+
+	err = api.Organization.UpdateClient(context.Background(), org.GetID(), client.GetClientID(), false)
+	require.NoError(t, err)
+
+	associatedClient, err = api.Organization.Client(context.Background(), org.GetID(), client.GetClientID())
+	require.NoError(t, err)
+	assert.False(t, associatedClient.GetUseForMemberAccess())
+
+	err = api.Organization.RemoveClient(context.Background(), org.GetID(), client.GetClientID())
+	require.NoError(t, err)
+
+	clients, err = api.Organization.Clients(context.Background(), org.GetID())
+	require.NoError(t, err)
+	assert.Len(t, clients.Clients, 0)
+}
+
 func TestOrganizationManager_ClientGrants(t *testing.T) {
 	configureHTTPTestRecordings(t)